Move Bootstrap feature specs for 'Download on Demand' under spec/features/webui/repositories_spec.rb
